requirements.txt
aws-cdk.core==1.141.0
aws-cdk.aws-cognito==1.141.0
aws-cdk.aws-lambda==1.141.0
aws-cdk.aws-apigateway==1.141.0
fastapi==0.68.0
uvicorn==0.15.0
aws-lambda-powertools==2.37.0
boto3==1.34.87
mangum==0.17.0